Edited Main Scene, Assemble_game Scene
-Last Scene에 저장된 데이터가 있는데 '새로하기' 버튼을 누른 경우 경고창(new_start_alert 오브젝트)을 띄움. (FadeOut.cs)
-Assemble_game1 Scene, Assemble_game2 Scene에서 부품들을 제대로 고르지 못한 경우 다음으로 넘어가는 버튼의 interactable을 false로 설정. (Components_check.cs)

+++ PBL_01/Assets/Scripts/FadeOut.cs
@@ -7,11 +7,16 @@ using UnityEngine.SceneManagement;
 public class FadeOut : MonoBehaviour
 {
     public Image Panel;
+    public Image new_start_alert;
     public Button new_start_btn;
     public Button continue_btn;
 
     float time = 0f;
     float F_time = 2f;
+
+    private void Awake() {
+        new_start_alert.gameObject.SetActive(false);
+    }
 
     public void F_Out(string scene_name)
     {
@@ -36,22 +41,45 @@ public class FadeOut : MonoBehaviour
     }
 
     public  void Click_Start_Btn() {
-        PlayerPrefs.DeleteAll();
-        PlayerPrefs.SetInt("out", 1);
-        PlayerPrefs.SetInt("day", 0);
-        StartCoroutine(UntilPlayback(new_start_btn, "Prologue_Hospital"));
+        if(PlayerPrefs.GetString("Last_scene") != "") { // 이어하기 데이터가 있으면
+            new_start_alert.gameObject.SetActive(true);
+            new_start_btn.GetComponent<AudioSource>().Play();
+        }
+        else {
+            PlayerPrefs.DeleteAll();
+            PlayerPrefs.SetInt("out", 1);
+            PlayerPrefs.SetInt("day", 0);
+            StartCoroutine(UntilPlayback(new_start_btn, "Prologue_Hospital"));
+        }
+        
     }
 
     public void Click_Continue_Btn() {
         StartCoroutine(UntilPlayback(continue_btn, PlayerPrefs.GetString("Last_scene")));
     }
 
+    public void Click_new_start_yes() {
+        PlayerPrefs.DeleteAll();
+        PlayerPrefs.SetInt("out", 1);
+        PlayerPrefs.SetInt("day", 0);
+        StartCoroutine(UntilPlayback(new_start_btn, "Prologue_Hospital"));
+    }
+
+    public void Click_new_start_no() {        
+        StartCoroutine(Click_new_start_no(new_start_alert.transform.Find("No_Button").GetComponent<Button>()));
+    }
     
     IEnumerator UntilPlayback(Button obj, string scene_name)
     {
         obj.GetComponent<AudioSource>().Play();
         yield return new WaitUntil(() => !obj.GetComponent<AudioSource>().isPlaying);
         F_Out(scene_name);
+    }
+
+    IEnumerator Click_new_start_no(Button obj) {
+        obj.GetComponent<AudioSource>().Play();
+        yield return new WaitUntil(() => !obj.GetComponent<AudioSource>().isPlaying);
+        new_start_alert.gameObject.SetActive(false);
     }
 
 }
